wip
wip

wip

geometeries

disocvery no frame config

Bump SDK ver v0.16.0

Pass link config and logger to resource config

Update desc

Update to 0.16.0 in conan file
