+++ meta.json
@@ -10,6 +10,12 @@
       "markdown_link": "README.md#model-viamorbbecastra2",
       "model": "viam:orbbec:astra2",
       "short_description": "Orbbec Astra2 driver return images, depth maps and pointclouds"
+    },
+      {
+      "api": "rdk:service:discovery",
+      "markdown_link": "README.md#model-viamorbbecdiscovery",
+      "model": "viam:orbbec:discovery",
+      "short_description": "A discovery service to find connected Orbbec cameras"
     }
   ],
   "entrypoint": "orbbec-module",
